Move the batch buffer to Espresso directory
